edited the scores so that disconnecting resets the averages, and also a possible fix to the removebot and removeallbots

+++ neo/d3xp/Confetti_Timer.h
@@ -3,5 +3,5 @@
 
 #include "Game_local.h"
 
-	extern float tickRate;			// Timer manipulator
+	extern int tickRate;			// Timer manipulator
 #endif
